Move backend data storage to dedicated module
