+++ mock-backend/mocks/fakeData/tagGenerators.ts
@@ -1,0 +1,5 @@
+import { faker } from '@faker-js/faker';
+
+export function makeRandomTags() {
+    return Array.from(Array(20), (value, key) => faker.company.bsNoun());
+};
