--- mock-backend/mocks/fakeData/tags.ts
@@ -1,3 +1,0 @@
-import { faker } from '@faker-js/faker';
-
-export const TAGS = Array.from(Array(20), (value, key) => faker.company.bsNoun());
