+++ mock-backend/mocks/fakeData/userGenerators.ts
@@ -1,14 +1,6 @@
 import { faker } from '@faker-js/faker';
 
-import { UserDTO, CurrentUserDTO } from '../models/user';
-
-export let CURRENT = makeRandomFakeCurrentUserDTO(0);
-export const USERS = [makeUserFromCurrentUser(CURRENT), ...Array.from(Array(20), (value, key) => makeRandomFakeUserDTO(key + 1))];
-
-export function updateCurrentUser(user: CurrentUserDTO) {
-	CURRENT = user;
-	USERS[0] = makeUserFromCurrentUser(user);
-}
+import { CurrentUserDTO, UserDTO } from '../models/user';
 
 
 export function makeRandomFakeCurrentUserDTO(id: number): CurrentUserDTO {
@@ -29,7 +21,7 @@ export function makeRandomFakeCurrentUserDTO(id: number): CurrentUserDTO {
     registrationDate: faker.date.past(),
     contactEmailAddress: faker.internet.email(firstName, lastName),
     isContactEmailAddressPublic: faker.datatype.boolean(),
-    phoneNumber: faker.phone.phoneNumber(),
+    phoneNumber: faker.phone.phoneNumber()
   };
 }
 
@@ -39,14 +31,14 @@ export function makeRandomFakeUserDTO(id: number): UserDTO {
   return makeUserFromCurrentUser(currentUser);
 }
 
-function makeUserFromCurrentUser(currentUser: CurrentUserDTO) {
-  const user = {...currentUser}  as UserDTO & Partial<CurrentUserDTO>;
+export function makeUserFromCurrentUser(currentUser: CurrentUserDTO) {
+  const user = { ...currentUser } as UserDTO & Partial<CurrentUserDTO>;
 
   delete user.phoneNumber;
   delete user.firstName;
   delete user.lastName;
-  if(!user.isContactEmailAddressPublic) {
-      delete user.contactEmailAddress;
+  if (!user.isContactEmailAddressPublic) {
+    delete user.contactEmailAddress;
   }
   delete user.isContactEmailAddressPublic;
